+++ app/uk/gov/hmrc/disguisedremunerationfrontend/data/Employer.scala
@@ -16,10 +16,4 @@
 
 package uk.gov.hmrc.disguisedremunerationfrontend.data
 
-import play.api.libs.json.{Format, Json}
-
 case class Employer(name: String, paye: Paye)
-
-object Employer {
-  implicit val EmployerFormatter: Format[Employer] = Json.format[Employer]
-}
