--- app/uk/gov/hmrc/disguisedremunerationfrontend/data/EnumUtils.scala
@@ -1,48 +1,0 @@
-/*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.disguisedremunerationfrontend.data
-
-import play.api.libs.json._
-
-/**
-  * Utility class for creating json formatters for enumerations.
-  */
-object EnumUtils {
-
-  def enumReads[E <: Enumeration](enum: E): Reads[E#Value] = new Reads[E#Value] {
-    def reads(json: JsValue): JsResult[E#Value] = json match {
-      case JsString(s) =>
-        try {
-          JsSuccess(enum.withName(s))
-        } catch {
-          case _: NoSuchElementException =>
-            JsError(s"Enumeration expected of type: '${enum.getClass}'," ++
-              s" but it does not appear to contain the value: '$s'")
-        }
-      case _ => JsError("String value expected")
-    }
-  }
-
-  implicit def enumWrites[E <: Enumeration]: Writes[E#Value] = new Writes[E#Value] {
-    def writes(v: E#Value): JsValue = JsString(v.toString)
-  }
-
-  implicit def enumFormat[E <: Enumeration](enum: E): Format[E#Value] = {
-    Format(enumReads(enum), enumWrites)
-  }
-
-}
